Style: 키워드 생성 버튼 위치를 그룹 목록으로 이동
